Add support for writing the FET430's firmware in uif-bsl.
This patch adds support to the uif-bsl driver for erasing the FET430's
memory, and writing to it. It also adds graceful ignoring of halt and
reset requests, which the "prog" command tries to get it to do.

I've tried this with my UIF and it appears to work fine.  It no longer
complains when I present it with an MSP430F247 after updating its
firmware :)

Be very careful!  I do *not* provide any guarantee that this won't
brick your UIF (which would require a second MSP430 programmer to
program the first...).