initial version of flash_bsl driver
This driver allows mspdebug to talk to the bootloader on a flash MSP430
via a serial port. DTR and RTS lines are used to control the TEST and
RST# pins on the MSP430, respectively. Currently, read and program are
supported. However, read is fairly useless because the driver currently
does not support any password other than the default. Thus it must
mass-erase the chip to gain access.

Read that again. USING THIS DRIVER WILL MASS ERASE YOUR CHIP IMMEDIATELY,
WITHOUT YOU ASKING FOR IT.