Add --force-reset option
Without this option MSP430F5526 fails to reset its
registers properly. When the SP register points to
a random address and we try to set a breakpoint with
gdb, it will try to read all memory locations from
SP up to the top of the RAM searching for the return
address in the current stack frame.

On MSP430F47173 initialization fails when not using
reset (command C_IDENT1 failed) and we also need
to fall back to using try_open with send_reset = 1.

Signed-off-by: Ionut Nicu <ionut.nicu@mindbit.ro>