Facelift MSPDEBUG_TILIB_PATH handling
Yes, I primarily just want to get rid of strcat (which is checked and
appears to be OK, but still, it's 2018).

Also bring behaviour in line with manual and common expectations:

- In absence of MSPDEBUG_TILIB_PATH, the dynamic linker search
  path is processed
- If MSPDEBUG_TILIB_PATH is specified, that directory is searched
- The original behaviour in case of an error is graceful fall-through
  to letting the dynamic linker do its deed
- The behaviour is changed such that if MSPDEBUG_TILIB_PATH is specified,
  only that path is being considered.

This is in line with both manual and common sense (otherwise one
could have just used LD_LIBRARY_PATH).